+++ az-flow-trigger-dependency-type/kafka-event-trigger/src/test/trigger/kafka/KafkaDepInstanceCollectionTest.java
